:truck: refactor: sepearte all of domains for less complexity (#13)
