+++ src/domain/family-member/dto/request/create-family-member.dto.ts
@@ -8,5 +8,9 @@ export class CreateFamilyMemberDto {
 
   @IsNotEmpty()
   @IsNumber()
+  familyId: Long;
+
+  @IsNotEmpty()
+  @IsNumber()
   role: number;
 }
