+++ src/domain/family-member/dto/response/response-family-member.dto.ts
@@ -1,0 +1,13 @@
+import { Post } from 'src/infra/entities';
+
+export class ResponseFamilyMemberDto {
+  familyMemberId: number;
+  familyId: number;
+  userId: number;
+
+  role: number;
+  pokeCount: number;
+  talkCount: number;
+
+  posts: Post[];
+}
