+++ src/domain/family-member/index.ts
@@ -1,0 +1,3 @@
+export * from './dto/request/create-family-member.dto';
+export * from './dto/request/update-family-member.dto';
+export * from './dto/response/response-family-member.dto';
