+++ src/domain/family-schedule/dto/response/response-family-schedule.dto.ts
