+++ src/domain/family-schedule/family-schedule.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreateFamilyScheduleDto, UpdateFamilyScheduleDto } from '../dto';
+import { CreateFamilyScheduleDto, UpdateFamilyScheduleDto } from '../family/dto';
 
 @Injectable()
 export class FamilyScheduleService {
