--- src/domain/family/controller/index.ts
@@ -1,7 +1,0 @@
-export * from './album.controller';
-export * from './family.controller';
-export * from './family-member.controller';
-export * from './family-schedule.controller';
-export * from './individual-schedule.controller';
-export * from './photo.controller';
-export * from './post.controller';
