--- src/domain/family/dto/family-member/index.ts
@@ -1,3 +1,0 @@
-export * from './request/create-family-member.dto';
-export * from './request/update-family-member.dto';
-export * from './response/response-family-member.dto';
