--- src/domain/family/dto/family-member/response/response-family-member.dto.ts
@@ -1,9 +1,0 @@
-export class ResponseFamilyMemberDto {
-  id: number;
-  name: string;
-  age: number;
-  kinship: string;
-  familyId: number;
-  createdAt: Date;
-  updatedAt: Date;
-}
