--- src/domain/family/dto/family/index.ts
@@ -1,3 +1,0 @@
-export * from './request/create-family.dto';
-export * from './request/update-family.dto';
-export * from './response/response-family.dto';
