+++ src/domain/family/dto/index.ts
@@ -1,14 +1,3 @@
-export * from './album/create-album.dto';
-export * from './album/update-album.dto';
-export * from './family/request/create-family.dto';
-export * from './family/request/update-family.dto';
-export * from './family-member/request/create-family-member.dto';
-export * from './family-member/request/update-family-member.dto';
-export * from './family-schedule/request/create-family-schedule.dto';
-export * from './family-schedule/request/update-family-schedule.dto';
-export * from './individual-schedule/create-individual-schedule.dto';
-export * from './individual-schedule/update-individual-schedule.dto';
-export * from './photo/request/create-photo.dto';
-export * from './photo/request/update-photo.dto';
-export * from './post/create-post.dto';
-export * from './post/update-post.dto';
+export * from './request/create-family.dto';
+export * from './request/update-family.dto';
+export * from './response/response-family.dto';
