--- src/domain/family/dto/individual-schedule/create-individual-schedule.dto.ts
@@ -1,1 +1,0 @@
-export class CreateIndividualScheduleDto {}
