--- src/domain/family/dto/individual-schedule/update-individual-schedule.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateIndividualScheduleDto } from './create-individual-schedule.dto';
-
-export class UpdateIndividualScheduleDto extends PartialType(
-  CreateIndividualScheduleDto,
-) {}
