--- src/domain/family/service/index.ts
@@ -1,7 +1,0 @@
-export * from './album.service';
-export * from './family.service';
-export * from './family-member.service';
-export * from './family-schedule.service';
-export * from './individual-schedule.service';
-export * from './photo.service';
-export * from './post.service';
