+++ src/domain/photo/dto/request/create-photo.dto.ts
