+++ src/domain/photo/dto/request/update-photo.dto.ts
