+++ src/domain/photo/dto/response/response-photo.dto.ts
