+++ src/domain/photo/photo.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreatePhotoDto, UpdatePhotoDto } from '../dto';
+import { CreatePhotoDto, UpdatePhotoDto } from '../family/dto';
 
 @Injectable()
 export class PhotoService {
