+++ src/test/e2e/photo.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { PhotoController } from '../../domain/family/controller/photo.controller';
-import { PhotoService } from '../../domain/family/service/photo.service';
+import { PhotoController } from '../../domain/photo/photo.controller';
+import { PhotoService } from '../../domain/photo/photo.service';
 
 describe('PhotoController', () => {
   let controller: PhotoController;
